// src/interfaces/ProfileData.ts
interface IProfileData {
  id?: number;
  first_name?: string;
  firstName?: string;
  lastName?: string;
  fullName?: string;
  userName?: string;
  email?: string;
  age?: number;
  status?: string;
  industry?: string;
  organization?: string;
  role?: string;
  city?: string;
  country?: string;
  website?: string;
  interest?: string;
  youtubeLink?: string;
  linkedinLink?: string;
  instagramLink?: string;
  facebookLink?: string;
  trainer?: string;
  short_bio?: string;
  educational_qualification?: string;
  trainer_experience?: number;
  areas_of_expertise?: string;
  sectors_of_cater?: string;
  profile_photo?: string;
  profile_video?: string;
  pdf_profile?: string;
  about_me?: string;
}

export default IProfileData;